Jmol.___JmolVersion="14.7.0_2016.06.22"
new feature: settable chain string using {atomset}.chain = "xxx"

bug fix: compare(a,b,"isomer") does not detect ENANTIOMER (broken in 14.5.5)
bug fix: missing error trap for unsettable property setting
bug fix: load models {0 0 1} ... fails in script compiler




Former-commit-id: 82f7e41fd1899a565bbc886f10419eddcd053d20

+++ Jmol/src/org/jmol/viewer/JmolStateCreator.java
@@ -27,9 +27,9 @@ public abstract class JmolStateCreator {
 
   abstract String getFunctionCalls(String selectedFunction);
 
-  abstract String getAtomicPropertyState(byte taintCoord, BS bsSelected);
+  abstract String getAtomicPropertyState(int taintCoord, BS bsSelected);
 
-  abstract void getAtomicPropertyStateBuffer(SB commands, byte type,
+  abstract void getAtomicPropertyStateBuffer(SB commands, int type,
                                     BS bs, String name, float[] data);
 
   abstract void undoMoveAction(int action, int n);
